Conflicts:
	tarjonta-app-angular/app/js/shared/koodisto.js
	tarjonta-app-angular/app/lib/tinymce/plugins/image/plugin.min.js
	tarjonta-app-angular/app/lib/tinymce/plugins/link/plugin.min.js
	tarjonta-app-angular/app/lib/tinymce/plugins/media/plugin.min.js
	tarjonta-app-angular/app/lib/tinymce/plugins/paste/plugin.min.js
	tarjonta-app-angular/app/lib/tinymce/plugins/table/plugin.min.js
	tarjonta-app-angular/app/lib/tinymce/themes/modern/theme.min.js
	tarjonta-app-angular/app/lib/tinymce/tinymce.min.js
	tarjonta-app-angular/app/partials/haku/edit/hakuEditControllers.js
	tarjonta-app/src/main/java/fi/vm/sade/tarjonta/ui/view/haku/ShowHakuView.java
	tarjonta-service/src/main/java/fi/vm/sade/tarjonta/service/impl/resources/v1/KoulutusResourceImplV1.java
	tarjonta-service/src/main/java/fi/vm/sade/tarjonta/service/impl/resources/v1/koulutus/validation/KoulutusValidator.java
	tarjonta-service/src/test/java/fi/vm/sade/tarjonta/service/impl/resources/v1/KoulutusResourceTest.java

+++ tarjonta-service/src/main/java/fi/vm/sade/tarjonta/service/impl/resources/v1/HakuResourceImplV1.java
@@ -228,7 +228,11 @@ public class HakuResourceImplV1 implements HakuV1Resource {
     @Override
     public ResultV1RDTO<HakuV1RDTO> createHaku(HakuV1RDTO haku) {
         LOG.info("createHaku() - {}", haku);
-        permissionChecker.checkCreateHaku();
+        if(haku.getOid()!=null) {
+            ResultV1RDTO<HakuV1RDTO> result = new ResultV1RDTO<HakuV1RDTO>();
+            createSystemErrorFromException(new RuntimeException("cannot create with predefeined oid"), result);
+            return result;
+        }
 
         return updateHaku(haku);
     }
@@ -237,7 +241,6 @@ public class HakuResourceImplV1 implements HakuV1Resource {
     @Override
     public ResultV1RDTO<HakuV1RDTO> updateHaku(HakuV1RDTO hakuDto) {
         LOG.info("updateHaku() - {}", hakuDto);
-        permissionChecker.checkUpdateHaku(hakuDto != null ? hakuDto.getOid() : null);
 
         ResultV1RDTO<HakuV1RDTO> result = new ResultV1RDTO<HakuV1RDTO>();
 
@@ -252,20 +255,38 @@ public class HakuResourceImplV1 implements HakuV1Resource {
 
             if (isNew) {
                 // Generate new OID for haku to be created
+                permissionChecker.checkCreateHakuWithOrgs(hakuDto.getTarjoajaOids());
+                
+
                 hakuDto.setOid(oidService.get(TarjontaOidType.HAKU));
                 LOG.info("updateHakue() - NEW haku! - oid ==> {}", hakuDto.getOid());
-            }
-
-            if (!isNew) {
+            } else {
                 LOG.info("updateHaku() - OLD haku - find by oid");
+                
+                final String oid = hakuDto.getOid();
 
-                // Check is haku exists
-                hakuToUpdate = hakuDAO.findByOid(hakuDto.getOid());
+                // Check haku exists
+                hakuToUpdate = hakuDAO.findByOid(oid);
+                
+                final TarjontaTila toTila = TarjontaTila.valueOf(hakuDto.getTila());
+
+                
+                //tilan muutos/testaus tilapalvelun kautta
+                final Tila tila = new Tila(Tyyppi.HAKU, toTila, oid);
+                if(!publication.isValidStatusChange(tila)){
+                    result.addError(ErrorV1RDTO.createValidationError("haku", "haku.tilamuutos.not.allowed", hakuDto.getOid()));
+                    result.setStatus(ResultV1RDTO.ResultStatus.ERROR);
+                    return result;
+                }
+                publication.updatePublicationStatus(Lists.newArrayList(tila));
+                
                 if (hakuToUpdate == null) {
                     result.addError(ErrorV1RDTO.createValidationError("haku", "haku.not.exists", hakuDto.getOid()));
                     result.setStatus(ResultV1RDTO.ResultStatus.ERROR);
                     return result;
                 }
+                //check permissions
+                permissionChecker.checkCreateHakuWithOrgs(hakuToUpdate.getTarjoajaOids());
             }
 
             LOG.info("updateHaku() - validate");
@@ -308,20 +329,29 @@ public class HakuResourceImplV1 implements HakuV1Resource {
     public ResultV1RDTO<Boolean> deleteHaku(String oid) {
         LOG.info("deleteHaku() oid={}", oid);
 
-        permissionChecker.checkRemoveHaku(oid);
 
         ResultV1RDTO<Boolean> result = new ResultV1RDTO<Boolean>();
         updateRightsInformation(result, null);
 
         Haku hakuToRemove = hakuDAO.findByOid(oid);
 
+        permissionChecker.checkRemoveHakuWithOrgs(hakuToRemove.getTarjoajaOids());
+        
         if (hakuToRemove != null) {
-            hakuDAO.remove(hakuToRemove);
-            result.setResult(true);
-            result.setStatus(ResultV1RDTO.ResultStatus.OK);
+            if (hakuToRemove.getHakukohdes().size() > 0) {
+                // Cannot delete HAKU with hakukohdes!
+                result.setResult(false);
+                result.setStatus(ResultStatus.ERROR);
+                result.addError(ErrorV1RDTO.createValidationError(null, "haku.delete.error.hasExistingHakukohdes"));
+            } else {
+                hakuDAO.remove(hakuToRemove);
+                result.setResult(true);
+                result.setStatus(ResultV1RDTO.ResultStatus.OK);
+            }
         } else {
             result.setResult(false);
             result.setStatus(ResultV1RDTO.ResultStatus.NOT_FOUND);
+            result.addError(ErrorV1RDTO.createValidationError(null, "haku.delete.error.notFound"));
         }
 
         return result;
@@ -433,25 +463,22 @@ public class HakuResourceImplV1 implements HakuV1Resource {
             result.addError(ErrorV1RDTO.createValidationError("nimi", "haku.validation.nimi.empty"));
         }
 
-        // Hakulomake URI  -OR- maxHakukohdes
-        if (isEmpty(haku.getHakulomakeUri())) {
-            // max hakuaikas cannot be empty
+        // If we are using systems application form - make sure the "maxHakukohdes" is set
+        if (haku.isJarjestelmanHakulomake()) {
             if (haku.getMaxHakukohdes() <= 0) {
                 result.addError(ErrorV1RDTO.createValidationError("maxHakukohdes", "haku.validation.maxHakukohdes.invalid"));
-            }
-        } else {
-//            // Cannot have this since we have hakulomakeUri
-//            if (haku.getMaxHakukohdes() > 0) {
-//                result.addError(ErrorV1RDTO.createValidationError("maxHakukohdes", "haku.validation.maxHakukohdes.invalid"));
-//            }
-
+            }            
+        }
+        
+        // Own hakulomake url? Verify it.
+        if (!isEmpty(haku.getHakulomakeUri())) {
             try {
                 URL url = new URL(haku.getHakulomakeUri());
             } catch (MalformedURLException ex) {
                 result.addError(ErrorV1RDTO.createValidationError("hakulomakeUri", "haku.validation.hakulomakeUri.invalid"));
             }
         }
-
+        
         // Must have at least one hakuaika
         if (haku.getHakuaikas() == null || haku.getHakuaikas().isEmpty()) {
             result.addError(ErrorV1RDTO.createValidationError("hakuaikas", "haku.validation.hakuaikas.empty"));
@@ -481,6 +508,11 @@ public class HakuResourceImplV1 implements HakuV1Resource {
         // TODO haku.getHakukausiVuosi() - verrataanko hakukausi / vuosi arvoihin?
         // TODO haku.getKoulutuksenAlkamisVuosi() - verrataanko hakukausi / vuosi arvoihin?
 
+        // Sijoittelu + system application form -> priority = true
+        if (haku.isSijoittelu() && isEmpty(haku.getHakulomakeUri())) {
+            haku.setUsePriority(true);
+        }        
+        
         if (result.hasErrors()) {
             result.setStatus(ResultV1RDTO.ResultStatus.ERROR);
